requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file marv/requirements.txt marv/requirements.in
#
alabaster==0.7.10         # via sphinx
babel==2.5.0              # via sphinx
bcrypt==3.1.3
certifi==2017.7.27.1      # via requests
cffi==1.10.0              # via bcrypt
chardet==3.0.4            # via requests
click==6.7
configparser==3.5.0
cython==0.26.1
docutils==0.14            # via sphinx
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.2
Flask==0.12.2             # via flask-cors, flask-sqlalchemy
idna==2.6                 # via requests
imagesize==0.7.1          # via sphinx
itsdangerous==0.24        # via flask
Jinja2==2.9.6             # via flask, sphinx
MarkupSafe==1.0           # via jinja2
oauthlib==2.0.3           # via requests-oauthlib
pycparser==2.18           # via cffi
Pygments==2.2.0           # via sphinx
PyJWT==1.5.3
pytz==2017.2              # via babel
requests-oauthlib==0.8.0
requests==2.18.4          # via requests-oauthlib, sphinx
Six==1.10.0               # via bcrypt, flask-cors, sphinx
snowballstemmer==1.2.1    # via sphinx
sphinx-paramlinks==0.3.4
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1  # via sphinx
SQLAlchemy==1.1.14        # via flask-sqlalchemy
typing==3.6.2             # via sphinx
urllib3==1.22             # via requests
Werkzeug==0.12.2          # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via sphinx
